requirements.txt
Django
djangorestframework
celery
redis
drf_yasg
psycopg2